requirements.txt
requests~=2.31.0
jsonschema~=4.21.1
pytest
allure-pytest~=2.13.5
